Warning fixes for Windows, FE-1757.

+++ FabricSpliceConversion.cpp
@@ -55,13 +55,13 @@ double getFloat64FromRTVal(FabricCore::RTVal rtVal)
   if(simpleData.type == FEC_RTVAL_SIMPLE_TYPE_UINT16)
     return simpleData.value.uint16;
   if(simpleData.type == FEC_RTVAL_SIMPLE_TYPE_UINT64)
-    return simpleData.value.uint64;
+    return double(simpleData.value.uint64);
   if(simpleData.type == FEC_RTVAL_SIMPLE_TYPE_SINT8)
     return simpleData.value.sint8;
   if(simpleData.type == FEC_RTVAL_SIMPLE_TYPE_SINT16)
     return simpleData.value.sint16;
   if(simpleData.type == FEC_RTVAL_SIMPLE_TYPE_SINT64)
-    return simpleData.value.sint64;
+    return double(simpleData.value.sint64);
   return DBL_MAX;
 }
 
@@ -906,10 +906,10 @@ void convertOutputLines(NurbsCurveList curveList, FabricCore::RTVal & rtVal)
           xsiPoints[vindex].GetY(),
           xsiPoints[vindex].GetZ(),
           1.0);
-        curveData[i].m_aKnots[j] = j;
+        curveData[i].m_aKnots[j] = double(j);
       }
       if(isClosed)
-        curveData[i].m_aKnots[nbVertices] = nbVertices;
+        curveData[i].m_aKnots[nbVertices] = double(nbVertices);
 
       coffset = voffset;
     }
